+++ src/main.cpp
@@ -5,25 +5,47 @@
 using namespace std;
 
 
+void run_simulation(int time_steps, int x_bound, int y_bound, int num_agents, int num_obstacles, int num_energies) {
+    BasicEnvironment environment(x_bound, y_bound, num_obstacles, num_energies);
+
+    // Creates all agents and inserts them into the environment
+    vector <BasicAgent*> agent_ptrs;
+    for (int i=0; i<num_agents; i++) {
+        string agent_name = "Agent " + to_string(i+1);
+        // Needs dynamic allocation to ensure objects survive out of the 'for' scope
+        BasicAgent* agent_ptr = new BasicAgent(agent_name, x_bound, y_bound);
+        environment.insert_agent(agent_ptr);
+        agent_ptrs.push_back(agent_ptr);
+    }
+    environment.visualise();
+
+    // For the required number of time steps, have all agents move randomly across the board
+    for (int i=0; i<time_steps; i++) {
+        for (BasicAgent* agent_ptr: agent_ptrs) {
+            agent_ptr->move_random();
+        }
+        environment.update();
+        environment.visualise();
+    }
+
+    //Kills off the agents, now that we're done with the simulation
+    for (BasicAgent* agent_ptr: agent_ptrs) {
+        delete agent_ptr;
+    }
+
+    cout << "Simulation complete!" << endl;
+}
+
+
 int main() {
+
     srand(time(NULL));
-    const int X_BOUND = 15;
-    const int Y_BOUND = 15;
+    int time_steps = 5;
+    int x_bound = 15;
+    int y_bound = 15;
+    int num_agents = 4;
+    int num_obstacles = 20;
+    int num_energies = 20;
 
-    BasicAgent agent_1("elo", X_BOUND, Y_BOUND);
-    BasicAgent agent_2("ahoy-hoy", X_BOUND, Y_BOUND);
-
-    BasicEnvironment environment(X_BOUND, Y_BOUND, 20, 20);
-    environment.state_objects();
-
-    environment.insert_agent(&agent_1);
-    environment.insert_agent(&agent_2);
-    environment.environment_visualise();
-
-    agent_1.move('X', 15);
-    agent_2.move('Y', 15);
-    environment.update();
-    environment.environment_visualise();
-
-    cout << "Success!" << endl;
+    run_simulation(time_steps, x_bound, y_bound, num_agents, num_obstacles, num_energies);
 }